Separating queries file into folders and files
