+++ client/storefront/src/graphql/mutations/login.ts
@@ -1,0 +1,10 @@
+import { gql } from "@apollo/client";
+
+export const LOGIN_USER = gql`
+  mutation Mutation($user: userLoginData) {
+    loginUser(user: $user) {
+      email
+      token
+    }
+  }
+`;
