+++ client/storefront/src/graphql/mutations/register.ts
@@ -1,0 +1,10 @@
+import { gql } from "@apollo/client";
+
+export const REGISTER_USER = gql`
+  mutation Mutation($user: userRegisterData) {
+    registerUser(user: $user) {
+      email
+      token
+    }
+  }
+`;
