+++ client/storefront/src/graphql/queries/getCategories.ts
@@ -1,0 +1,10 @@
+import { gql } from "@apollo/client";
+
+export const GET_CATEGORIES = gql`
+  query Query {
+    getCategories {
+      id
+      name
+    }
+  }
+`;
